+++ src/Business/Grand.Business.Core/Interfaces/Catalog/Products/IInventoryManageService.cs
@@ -19,7 +19,7 @@ public interface IInventoryManageService
     /// <param name="warehouseId">Warehouse ID (for multi-warehouse scenarios)</param>
     /// <param name="userId">User ID who made the change</param>
     /// <param name="attributes">Product attributes (for attribute-based inventory)</param>
-    Task UpdateStockProduct(Product product, bool mediator = true, bool trackInventory = false, int? previousStockQuantity = null, string warehouseId = null, string userId = null, IList<CustomAttribute> attributes = null);
+    Task UpdateStockProduct(Product product, bool mediator = true, bool trackInventory = false, double? previousStockQuantity = null, string warehouseId = null, string userId = null, IList<CustomAttribute> attributes = null);
 
     /// <summary>
     ///     Adjust reserved inventory
@@ -28,7 +28,7 @@ public interface IInventoryManageService
     /// <param name="quantityToChange">Quantity to increase or decrease</param>
     /// <param name="attributes">Attributes</param>
     /// <param name="warehouseId">Warehouse ident</param>
-    Task AdjustReserved(Product product, int quantityToChange, IList<CustomAttribute> attributes = null,
+    Task AdjustReserved(Product product, double quantityToChange, IList<CustomAttribute> attributes = null,
         string warehouseId = "");
